// src/tcgdb/modules/tcgdb/pokemon/price/change/daily/dto/tcgdb.pokemon.price.change.daily.dto.ts
export class TCGdbPokemonPricesChangeDailyDTO {
    tcgdbPokemonPricesChangeDaily: TCGdbPokemonPriceChangeDailyDTO[];
}

export class TCGdbPokemonPriceChangeDailyDTO {
    tcgdbPokemonPriceChangeDailyId: string;
    tcgdbPokemonCardId: string;
    tcgdbPokemonPriceChangeDailyTCGPlayerId: number;
    tcgdbPokemonPriceChangeDailySetCode: string;
    tcgdbPokemonPriceChangeDailyCurrentLowPrice: number;
    tcgdbPokemonPriceChangeDailyPreviousLowPrice: number;
    tcgdbPokemonPriceChangeDailyLowPriceDifference: number;
    tcgdbPokemonPriceChangeDailyLowPricePercentage: number;
    tcgdbPokemonPriceChangeDailyCurrentMidPrice: number;
    tcgdbPokemonPriceChangeDailyPreviousMidPrice: number;
    tcgdbPokemonPriceChangeDailyMidPriceDifference: number;
    tcgdbPokemonPriceChangeDailyMidPricePercentage: number;
    tcgdbPokemonPriceChangeDailyCurrentHighPrice: number;
    tcgdbPokemonPriceChangeDailyPreviousHighPrice: number;
    tcgdbPokemonPriceChangeDailyHighPriceDifference: number;
    tcgdbPokemonPriceChangeDailyHighPricePercentage: number;
    tcgdbPokemonPriceChangeDailyCurrentMarketPrice: number;
    tcgdbPokemonPriceChangeDailyPreviousMarketPrice: number;
    tcgdbPokemonPriceChangeDailyMarketPriceDifference: number;
    tcgdbPokemonPriceChangeDailyMarketPricePercentage: number;
    tcgdbPokemonPriceChangeDailySubTypeName: string;
}

export class CreateTCGdbPokemonPriceChangeDailyDTO {
    tcgdbPokemonCardId: string;
    tcgdbPokemonPriceChangeDailyTCGPlayerId: number;
    tcgdbPokemonPriceChangeDailySetCode: string;
    tcgdbPokemonPriceChangeDailyCurrentLowPrice: number;
    tcgdbPokemonPriceChangeDailyPreviousLowPrice: number;
    tcgdbPokemonPriceChangeDailyLowPriceDifference: number;
    tcgdbPokemonPriceChangeDailyLowPricePercentage: number;
    tcgdbPokemonPriceChangeDailyCurrentMidPrice: number;
    tcgdbPokemonPriceChangeDailyPreviousMidPrice: number;
    tcgdbPokemonPriceChangeDailyMidPriceDifference: number;
    tcgdbPokemonPriceChangeDailyMidPricePercentage: number;
    tcgdbPokemonPriceChangeDailyCurrentHighPrice: number;
    tcgdbPokemonPriceChangeDailyPreviousHighPrice: number;
    tcgdbPokemonPriceChangeDailyHighPriceDifference: number;
    tcgdbPokemonPriceChangeDailyHighPricePercentage: number;
    tcgdbPokemonPriceChangeDailyCurrentMarketPrice: number;
    tcgdbPokemonPriceChangeDailyPreviousMarketPrice: number;
    tcgdbPokemonPriceChangeDailyMarketPriceDifference: number;
    tcgdbPokemonPriceChangeDailyMarketPricePercentage: number;
    tcgdbPokemonPriceChangeDailySubTypeName: string;
}